refactor(cli): extract signals command family into cli/commands/signals.ts
