+++ apps/axctl/src/cli/index.ts
@@ -113,14 +113,13 @@ import {
     type GroupByKey,
 } from "../metrics/aggregates.ts";
 import { fetchSessionDurabilityDetail } from "../metrics/reverted-commits.ts";
-import { cleanSessionId, formatSessionMetrics, SESSION_METRICS_LEGEND } from "../metrics/util.ts";
-import { SIGNAL_CATALOG, findSignal, runRelationSignal } from "../metrics/catalog.ts";
-import type { CascadeEdge } from "../metrics/fragility-cascade.ts";
+import { formatSessionMetrics, SESSION_METRICS_LEGEND } from "../metrics/util.ts";
 import { fetchLocSummary, type LocSummary, type LocSelector } from "../dashboard/loc-query.ts";
 import { renderSessionMarkdown, renderSessionJson } from "./session-show-format.ts";
 import { renderCompareTable, renderCompareJson } from "./session-compare-format.ts";
 import { cmdDaemon, cmdDoctor, cmdInstall, cmdSetup, cmdUninstall } from "./install.ts";
 import { insightsCommand, reportCommand, timelineCommand, reportRuntime } from "./commands/report.ts";
+import { signalsCommand, signalsRuntime } from "./commands/signals.ts";
 import type { RuntimeManifest } from "./commands/manifest.ts";
 import { resolvePwdRepository } from "../pwd.ts";
 import { detectStaleness } from "@ax/lib/transcript-staleness";
@@ -3590,70 +3589,6 @@ const sessionsCommand = Command.make("sessions").pipe(
     ]),
 );
 
-const formatCascadeEdges = (edges: readonly CascadeEdge[], descriptor: { label: string }): string => {
-    if (edges.length === 0) return `${descriptor.label}: no edges (no reverted-commit files have downstream fixers).`;
-    const lines: string[] = [];
-    lines.push(`${descriptor.label} (${edges.length} edge${edges.length === 1 ? "" : "s"}):`);
-    for (const e of edges) {
-        lines.push(`  ${cleanSessionId(e.origin)} → ${cleanSessionId(e.downstream)}  (weight ${e.weight})`);
-    }
-    return lines.join("\n");
-};
-
-const cmdSignalsList = Effect.sync(() => {
-    const lines = SIGNAL_CATALOG.map(
-        (s) => `${s.id}  [${s.kind}]  ${s.label} - ${s.description}`,
-    );
-    console.log(lines.join("\n"));
-});
-
-const cmdSignalsShow = (input: { readonly id: string; readonly limit: number; readonly json: boolean }) =>
-    Effect.gen(function* () {
-        const descriptor = findSignal(input.id);
-        if (descriptor === undefined) {
-            const ids = SIGNAL_CATALOG.map((s) => s.id).join(", ");
-            process.stderr.write(`axctl signals show: unknown signal "${input.id}". Valid ids: ${ids}\n`);
-            process.exit(2);
-            return;
-        }
-        if (descriptor.kind === "aggregate") {
-            console.log("aggregate rendering is a later wave");
-            return;
-        }
-        const all = yield* runRelationSignal(descriptor.id);
-        const sorted = [...all].sort((a, b) => b.weight - a.weight).slice(0, input.limit);
-        if (input.json) {
-            console.log(prettyPrint(sorted));
-            return;
-        }
-        console.log(formatCascadeEdges(sorted, descriptor));
-    });
-
-const signalsListCommand = Command.make("list", {}, () => cmdSignalsList).pipe(
-    Command.withDescription("Print the signal catalog: one line per signal (id [kind] label - description)."),
-);
-
-const signalsShowCommand = Command.make(
-    "show",
-    {
-        id: Argument.string("id"),
-        limit: positiveLimit(30),
-        json: jsonFlag,
-    },
-    ({ id, limit, json }) => cmdSignalsShow({ id, limit, json }),
-).pipe(Command.withDescription(
-    "Render a signal by id. Relation signals (e.g. fragility_cascade) print origin → downstream edges "
-    + "sorted by weight (top --limit, default 30). --json for machine output.",
-));
-
-const signalsCommand = Command.make("signals").pipe(
-    Command.withDescription("Signal catalog: list (browse) + show <id> (render a cross-session signal, e.g. fragility_cascade)"),
-    Command.withSubcommands([
-        signalsListCommand,
-        signalsShowCommand,
-    ]),
-);
-
 const improveCommand = Command.make("improve").pipe(
     Command.withDescription("Experiment loop: rank proposals (recommend), accept (emit task brief or scaffold + dispatch subagent), lint grounded agent files, track verdicts at +3/+10/+30 sessions after accept."),
     Command.withSubcommands([
@@ -5557,7 +5492,6 @@ const LEGACY_RUNTIME: RuntimeManifest = {
     "derive-intents": "db",
     classifiers: "db",
     sessions: "db",
-    signals: "db",
     improve: "db",
     retro: "db",
     costs: "db",
@@ -5579,6 +5513,7 @@ const LEGACY_RUNTIME: RuntimeManifest = {
 export const RUNTIME_BY_COMMAND: RuntimeManifest = {
     ...LEGACY_RUNTIME,
     ...reportRuntime,
+    ...signalsRuntime,
 };
 
 // Commands whose handlers reach into SurrealClient via AppLayer (or the
